#11 [refactor] 카톡 로그인 공백 수정
