+++ src/main/java/com/moddy/server/controller/auth/dto/request/SocialLoginRequest.java
@@ -1,7 +1,6 @@
-package com.moddy.server.request;
+package com.moddy.server.controller.auth.dto.request;
 
 public record SocialLoginRequest(
         String code
 ) {
-
 }
