+++ src/main/java/com/moddy/server/controller/auth/dto/response/SocialLoginResponse.java
@@ -1,4 +1,4 @@
-package com.moddy.server.response;
+package com.moddy.server.controller.auth.dto.response;
 
 
 public record SocialLoginResponse(
