+++ src/main/java/com/moddy/server/external/kakao/SocialPlatform.java
@@ -7,8 +7,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
 public enum SocialPlatform {
-    KAKAO("카카오"),
-    ;
+    KAKAO("카카오");
 
     private final String value;
 }
