+++ src/main/java/com/moddy/server/external/kakao/service/SocialService.java
@@ -1,8 +1,8 @@
-package com.moddy.server.service;
+package com.moddy.server.external.kakao.service;
 
 
-import com.moddy.server.request.SocialLoginRequest;
-import com.moddy.server.response.SocialLoginResponse;
+import com.moddy.server.controller.auth.dto.request.SocialLoginRequest;
+import com.moddy.server.controller.auth.dto.response.SocialLoginResponse;
 
 public abstract class SocialService {
     public abstract SocialLoginResponse login(SocialLoginRequest request);
